requirements.txt
flask
mysql-connector-python
mysqlclient
